Move tests to `quickwit-codegen-example` crate
