+++ quickwit/quickwit-codegen-example/Cargo.toml
@@ -1,0 +1,27 @@
+[package]
+name = "quickwit-codegen-example"
+version = "0.4.0"
+authors = ["Quickwit, Inc. <hello@quickwit.io>"]
+edition = "2021"
+license = "AGPL-3.0-or-later"                           # For a commercial, license, contact hello@quickwit.io
+description = "Provides an example of code generation for a simple service."
+repository = "https://github.com/quickwit-oss/quickwit"
+homepage = "https://quickwit.io/"
+documentation = "https://quickwit.io/docs/"
+
+[dependencies]
+async-trait = { workspace = true }
+dyn-clone = { workspace = true }
+prost = { workspace = true }
+serde = { workspace = true }
+thiserror = { workspace = true }
+tokio = { workspace = true }
+tonic = { workspace = true }
+tower = { workspace = true }
+utoipa = { workspace = true }
+
+[dev-dependencies]
+mockall = { workspace = true }
+
+[build-dependencies]
+quickwit-codegen = { workspace = true }
