+++ quickwit/quickwit-codegen-example/src/error.rs
