+++ quickwit/quickwit-codegen-example/src/hello.proto
